refactor(datasets): isolate protocol metric expression helpers
